'processed' flag so nodes only run once.
There was a problem with unintended multiple executions of nodes.
Added a flag (self.processed) which is set to True when a node
runs, which ensures it only runs once. The flags for all nodes
are reset by ProcessingNetwork when it runs process() for the
network (at the beginning i.e. before it actually runs the
network processing).